This change is an attempt to resolve the long standing issue where executables made via 'nekotools boot <bytecode>' on Linux systems (really systems using ELF ABI) will be irreprably damaged if stripped by the 'strip' toolchain program.  The issue is described in detail here...
  https://github.com/HaxeFoundation/neko/issues/75

The solution here is to modify the Neko VM and the nekotools programs
to be more ELF-aware.  This new code is activated by compiling with
ABI_ELF (now the default for Linux builds).

The Neko VM will now attempt to find a '.nekobytecode' section in the
executable by rifling through the ELF section headers.  If such a
section is found, it is inspected, looking for the customary bytecode
signature at the end of the section.  If the section is not found, the
code will fall back to the old algorithm and search at the end of the
file.  (This algorithm was necessary since tools like strip may
actually place other sections after the bytecode at the end of the
file, even if the bytecode section was initially at the end of the
file.)

Additionally, the 'nekotools' executable now understands how to inject
bytecode into the '.nekobytecode' section, resizing that section as
appropriate.  This allows this data to survive strip attempts.

Below are some measured performance and executable size impacts of
these changes as measured on my Fedora Core 17, x86_64 box that uses
Xeon processors...

  perf : 'neko -version' execution slowed by 32% or ~4.4ms/execution on my 2.53Ghz Xeon
  perf : 'nekoc -help' execution slowed by 4% or ~13.8ms/execution on my 2.53Ghz Xeon
  size : neko vm grew 2,880 bytes or ~23% larger; still under 16KB
  size : nekotools boot executables can be stripped now, making them 2%-8% smaller on x86_64