Keep reference to k_db abstract and avoid early finalizer call
In the SQLite CFFI library, `connect` returns an abstract Neko value, of
`k_db` kind and with a finalizer set.  If the `k_result` request value
only keeps a pointer to the db struct, instead of the corresponding
`k_db` value, the finalizer will be called as soon as no other external
references remain to that connection, even if the request itself is
still reachable.

The issue could manifest in code like (Haxe):

	var rs = Sqlite.open("db.db").request("select * from tbl");
	trace(rs.length);