Fix build errors on Xcode 12: (#218)
libs/ui/ui.c:197:2: error: implicit declaration of function 'RunApplicationEventLoop' is invalid in C99 [-Werror,-Wimplicit-function-declaration]
   libs/ui/ui.c:217:2: error: implicit declaration of function 'QuitApplicationEventLoop' is invalid in C99 [-Werror,-Wimplicit-function-declaration]

These APIs seem to be (semi?) private.  They exist, but in the
public SDK they are only defined in CarbonEvents.h deep inside
HIToolbox.  I'm not sure if there is a better fix for this, but
this seems to compile OK, even if it is a little hacky.