Fix new aprutil build on windows
We now have to download and build the expat library separately, as it is
no longer bundled with aprutil.

This requires patching the aprutil cmakelists file, as cmake's findexpat
script is currently quite limited. See:
https://gitlab.kitware.com/cmake/cmake/-/issues/24422
It also requires setting the CMP0074 policy (available from cmake 3.12)
when building aprutil, so that EXPAT_ROOT is not ignored.