[cmake] Update static dependencies to fix build issues (#272)
* Update mariadb-connector-c to 3.3.4

Fixes #216

* Update apache, apr, and aprutil

This means we no longer need to use the trunk version for mac, since the
new releases have the desired patches for:
https://bz.apache.org/bugzilla/show_bug.cgi?id=64753

Also means we no longer need pcre1, since now apache can use pcre2.

Apache 2.4.55, apr 1.7.2, and aprutil 1.6.3.

* Update remaining dependencies

OpenSSL 1.1.1t, pcre2 10.42, SQLite 3.40.1, and MbedTLS 2.28.2

* Remove unneeded cmake scripts

* Use sha256 for zlib instead of md5

* Reuse zlib build for mariadbconnector

This allows mariadbconnector to reuse the static zlib build created for
zlib.ndll, or to use the system version otherwise on unix systems.

* Fix pcre linking

* Fix make cflags for pcre2 build on mac

* Fix and update cmake patches

Remove obsolete patches

Avoid patch error with semicolons
https://stackoverflow.com/questions/72098319/cmake-string-replace-removes-semi-colon

* Fix openssl linking for mariadbconnector

Should ensure static linking on all platforms if STATIC_OPENSSL is set.

* Cleanup openssl config

* Remove openssl config from aprutil build

These flags are not enough to build aprutil with openssl support,
so they have had no effect even for aprutil 1.5.4.

* Update cmake version to 3.14

This gives access to FindSQLite3.

* Fix new aprutil build on windows

We now have to download and build the expat library separately, as it is
no longer bundled with aprutil.

This requires patching the aprutil cmakelists file, as cmake's findexpat
script is currently quite limited. See:
https://gitlab.kitware.com/cmake/cmake/-/issues/24422
It also requires setting the CMP0074 policy (available from cmake 3.12)
when building aprutil, so that EXPAT_ROOT is not ignored.

* Fix apache static linking to pcre on windows

* Update changelog

* Fix azure pipeline step name

* Use add_compile_definitions

We can use this now with cmake 3.12+

* Avoid error when only Apache is set in STATIC_DEPS

If Apache was set in STATIC_DEPS, but not APR or APRutil, this would
cause a build error previously.

* Fix GNUInstallDirs warning

project() call should come first

* Clean up minor issues

* Remove call to find openssl

If we are not doing a static build, then there is no reason to know
anything about the openssl library. That will be handled by the
mariadbconnector build scripts.

* Fix "WARNING: OpenSSL version is too old"

* Fix openssl static linking

* Add version check for expat locating code

* Clean up cmake version checks

* Fix "no known rule" error for libssl.a

* Update CHANGES

* Fix changelog

* Restore cmake 3.10.2 support

* Add warning about CMake 3.12 on Windows

When building APRutil, we rely on EXPAT_ROOT for finding expat, which is
only available with CMP0074 on CMake 3.12+. This means building APRutil
on Windows will fail if CMake is older than 3.12.

* Remove redundant policy setting

This policy is from CMake 3.9, so it gets set automatically with:
cmake_minimum_required(VERSION 3.10.2)

* Update mbedtls to 2.28.3

---------

Co-authored-by: Simon Krajewski <simon@haxe.org>